fix(windows): use Windows NLS APIs for number formats
`RNLocalizeModule::getNumberFormatSettings` was using a Windows API to
format a number (1,000) and then extract formatting characters from there.
It was relying on the fact that the length should be greater than 5,
which isn't always the case as on Windows users can heavily customize
decimal number formatting by for instance removing the whole decimal
part of the displayed number. In this case, `getNumberFormatSettings`
was just crashing.

Instead of trying to get localization information based on the output of
the format method, this commit uses Windows NLS APIs (National Language
Support) to directly retrieves the settings.

In addition to fixing a potential crash, it also fixes encoding as some
languages use Unicode characters in their number formats.
For instance in fr-FR, "1,000" is "1 000,00" with a "," as a decimal
separator and " " (U+202F - Narrow non-breaking space) as thousand separator.